WIP refactoring for better errors in tokens * most code is disabled * codepoints track position on their own * tokenizer partial working (number & string tests work)

+++ shared/strings.lib/strings/Counter.h
@@ -1,12 +1,12 @@
 #pragma once
-#include <cinttypes>
+#include <cstddef>
 
 namespace strings {
 
 /// naively strong typed counter
 struct Counter {
     using This = Counter;
-    uint32_t v{};
+    size_t v{};
 
     constexpr bool operator==(This o) const noexcept { return v == o.v; }
     constexpr bool operator!=(This o) const noexcept { return v != o.v; }
